Skip pod deletion stage if PodDeletionFilter was not defined
Currently, the check for podDeletionEnabled is performed
only WaitForCompletion spec is not provided. In this case,
if pod deletion is disabled, node states will move directly
from UpgradeStateWaitForJobsRequired to UpgradeStateDrainRequired.
Otherwise, when the waiting for completion stage is finished,
the pods unconditionally move to UpgradeStatePodDeletionRequired.
To address the issue, adding a check for podDeletionEnabled when
processing the UpgradeStatePodDeletionRequired pods to move them
to the drain stage if needed.

Signed-off-by: amaslennikov <amaslennikov@nvidia.com>