Add support to orphaned driver pod
- Add orphaned driver pods in the state built
- Ignore not scheduled pod in the state built
- Support upgrade-requested annotation to force moving to
  upgrade-required state

With this functionality, upgrading from a DS to a new one
is possible, with the following assumptions:

- New DS should have Node Anti Affinity to prevent scheduling
new driver pods where old still run.
- The old DS should be deleted by Operator with DeletePropagationOrphan
option to keep the old driver pods running until the upgrade flow
replaces them.

In addition, it will also be possible to only detach the pod from
DaemonSet to migrate to new DaemonSet.

Signed-off-by: Fred Rolland <frolland@nvidia.com>